src/python/projects/ai_models/opencv_face_detection/requirements.txt
mlflow
virtualenv
python-dotenv
